+++ src/components/molecules/Popup/index.ts
@@ -2,10 +2,7 @@ import "./styles.css";
 
 import { PopupComponent } from "../../../types";
 
-import {
-  buildMbxStandardComponent,
-  buildMobrixUiStandardComponent,
-} from "../../../tools";
+import { buildMbxStandardComponent } from "../../../tools";
 
 import popupComponent from "./component";